# Optimize slow tests in activerecord