--- a/activerecord/test/models/paragraph.rb
+++ b/activerecord/test/models/paragraph.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class Paragraph < ActiveRecord::Base
+  belongs_to :book
+end