Follow up to https://github.com/rails/rails/pull/47499.

Was able to substantially speedup active record tests even further (this is probably the last time) by working out through the slowest tests.

The main offenders are schema dumping and schema cache tests, which dump/load a lot of tables from the [schema.rb file](https://github.com/rails/rails/blob/main/activerecord/test/schema/schema.rb) connected to `:arunit` connections, where we can work with a much smaller `:arunit2` connections' tables.

`$ CI=1 SEED=1 rake test:postgresql`

### Timings
Before: `Finished in 199.307319s, 44.0476 runs/s, 170.5808 assertions/s.`
After: `Finished in 162.104721s, 54.1563 runs/s, 160.5629 assertions/s.` 🔥 

### Slowest tests
Used https://github.com/y-yagi/minitest-test_profile for detection.

#### Before
```
Top 30 slowest tests (79.700740 seconds, 40.2% of total time):
PooledConnectionsTest#test_pooled_connection_checkin_one
  20.23 seconds
ActiveRecord::TooManyOrTest#test_too_many_or
  6.80 seconds
EagerLoadingTooManyIdsTest#test_eager_loading_too_many_ids
  5.12 seconds
PessimisticLockingTest#test_no_locks_no_wait
  5.03 seconds
ActiveRecord::Encryption::ConcurrencyTest#test_models_can_be_encrypted_and_decrypted_in_different_threads_concurrently
  3.25 seconds
SecurePasswordTest#test_authenticate_by_takes_the_same_amount_of_time_regardless_of_whether_record_is_found
  2.91 seconds
SchemaDumperTest#test_schema_dump_with_table_name_prefix_and_ignoring_tables
  2.19 seconds
ActiveRecord::DatabaseTasksDumpSchemaTest#test_db_dir_ignored_if_included_in_schema_dump
  2.17 seconds
ActiveRecord::DatabaseTasksDumpSchemaTest#test_ensure_db_dir
  2.16 seconds
SchemaDumperTest#test_schema_dump_keeps_id_false_when_id_is_false_and_unique_not_null_column_added
  2.14 seconds
SchemaDumperTest#test_schema_dump_with_regexp_ignored_table
  2.14 seconds
ActiveRecord::ConnectionAdapters::SchemaCacheTest#test_yaml_dump_and_load
  1.92 seconds
ActiveRecord::ConnectionAdapters::SchemaCacheTest#test_when_lazily_load_schema_cache_is_set_cache_is_lazily_populated_when_est_connection
  1.71 seconds
EagerLoadingTooManyIdsTest#test_preloading_too_many_ids
  1.66 seconds
PostgresqlHstoreTest#test_disable_enable_hstore
  1.63 seconds
ActiveRecord::ConnectionAdapters::SchemaCacheTest#test_yaml_dump_and_load_with_gzip
  1.61 seconds
ActiveRecord::ConnectionAdapters::SchemaCacheTest#test_cache_path_can_be_in_directory
  1.44 seconds
ActiveRecord::ConnectionAdapters::SchemaCacheTest#test_marshal_dump_and_load_with_ignored_tables
  1.42 seconds
ActiveRecord::ConnectionAdapters::SchemaCacheTest#test_marshal_dump_and_load_with_gzip
  1.39 seconds
ActiveRecord::ConnectionAdapters::SchemaCacheTest#test_marshal_dump_and_load_via_disk
  1.39 seconds
ActiveRecord::DatabaseTasksDumpSchemaCacheTest#test_dump_schema_cache
  1.37 seconds
ActiveRecord::ConnectionAdapters::ConnectionPoolThreadTest#test_disconnect_and_clear_reloadable_connections_attempt_to_wait_for_threads_to_return_their_conns
  1.26 seconds
ActiveRecord::BindParameterTest#test_too_many_binds_with_query_cache
  1.19 seconds
ActiveRecord::ConnectionAdapters::ConnectionPoolThreadTest#test_checkout_fairness
  1.19 seconds
ActiveRecord::PostgresqlTransactionNestedTest#test_deadlock_raises_Deadlocked_inside_nested_SavepointTransaction
  1.17 seconds
ActiveRecord::PostgresqlTransactionNestedTest#test_deadlock_inside_nested_SavepointTransaction_is_recoverable
  1.13 seconds
ActiveRecord::ConnectionAdapters::PostgreSQLAdapter::StatementPoolTest#test_cache_is_per_pid
  1.11 seconds
ActiveRecord::PostgresqlTransactionTest#test_raises_Deadlocked_when_a_deadlock_is_encountered
  1.06 seconds
ActiveRecord::DatabaseTasksTruncateAllTest#test_truncate_tables
  0.99 seconds
ActiveRecord::BindParameterTest#test_too_many_binds
  0.91 seconds
```

#### After
```
Top 30 slowest tests (53.339880 seconds, 33.1% of total time):
PooledConnectionsTest#test_pooled_connection_checkin_one
  20.24 seconds
EagerLoadingTooManyIdsTest#test_eager_loading_too_many_ids
  5.58 seconds
PessimisticLockingTest#test_no_locks_no_wait
  5.02 seconds
SecurePasswordTest#test_authenticate_by_takes_the_same_amount_of_time_regardless_of_whether_record_is_found
  2.96 seconds
SchemaDumperTest#test_schema_dump_keeps_id_false_when_id_is_false_and_unique_not_null_column_added
  2.18 seconds
PostgresqlHstoreTest#test_disable_enable_hstore
  1.62 seconds
EagerLoadingTooManyIdsTest#test_preloading_too_many_ids
  1.51 seconds
ActiveRecord::BindParameterTest#test_too_many_binds_with_query_cache
  1.36 seconds
ActiveRecord::ConnectionAdapters::ConnectionPoolThreadTest#test_disconnect_and_clear_reloadable_connections_attempt_to_wait_for_threads_to_return_their_conns
  1.25 seconds
ActiveRecord::ConnectionAdapters::ConnectionPoolThreadTest#test_checkout_fairness
  1.20 seconds
ActiveRecord::BindParameterTest#test_too_many_binds
  1.11 seconds
ActiveRecord::PostgresqlTransactionNestedTest#test_deadlock_inside_nested_SavepointTransaction_is_recoverable
  1.04 seconds
ActiveRecord::PostgresqlTransactionNestedTest#test_deadlock_raises_Deadlocked_inside_nested_SavepointTransaction
  1.04 seconds
ActiveRecord::PostgresqlTransactionTest#test_raises_Deadlocked_when_a_deadlock_is_encountered
  1.03 seconds
ConcurrentTransactionTest#test_transaction_isolation__read_committed
  0.58 seconds
ActiveRecord::PostgresqlTransactionTest#test_raises_QueryCanceled_when_canceling_statement_due_to_user_request
  0.53 seconds
ActiveRecord::Encryption::KeyGeneratorTest#test_derive_key_derives_a_key_with_from_the_provided_password_with_the_cipher_key_length_by_default
  0.53 seconds
ActiveRecord::PostgresqlTransactionTest#test_raises_Interrupt_when_canceling_statement_via_interrupt
  0.51 seconds
ActiveRecord::Encryption::DerivedSecretKeyProviderTest#test_will_derive_a_key_with_the_right_length_from_the_given_password
  0.42 seconds
EagerLoadPolyAssocsTest#test_include_query
  0.40 seconds
ActiveRecord::ConnectionAdapters::ConnectionHandlerTest#test_forked_child_recovers_from_disconnected_parent
  0.37 seconds
ActiveRecord::Encryption::KeyProviderTest#test_work_with_multiple_keys_when_config.store_key_references_is_false
  0.32 seconds
ActiveRecord::Encryption::SchemeTest#test_validates_config_options_when_declaring_encrypted_attributes
  0.32 seconds
ActiveRecord::Encryption::DerivedSecretKeyProviderTest#test_work_with_multiple_keys_when_config.store_key_references_is_true
  0.32 seconds
ActiveRecord::Encryption::KeyProviderTest#test_when_the_message_does_not_contain_any_key_reference,_it_returns_all_the_keys
  0.32 seconds
ActiveRecord::Encryption::KeyProviderTest#test_when_store_key_references_is_true,_the_encryption_key_contains_a_reference_to_the_key_itself
  0.32 seconds
ActiveRecord::Encryption::KeyProviderTest#test_when_the_message_to_decrypt_contains_a_reference_to_the_key_id,_it_will_return_an_array_only_with_that_message
  0.32 seconds
ActiveRecord::Encryption::KeyProviderTest#test_work_with_multiple_keys_when_config.store_key_references_is_true
  0.32 seconds
ActiveRecord::Encryption::KeyProviderTest#test_when_store_key_references_is_false,_the_encryption_key_contains_a_reference_to_the_key_itself
  0.32 seconds
ActiveRecord::Encryption::KeyProviderTest#test_serves_the_last_key_for_encrypting
  0.32 seconds
```

Note: There is an offender with 20s runtime at the top (😱), so someone else with the skills can try to get rid of that, if they wish.