@fatkodima 2023-02-27
Solved CI failure. The problem was that for sqlite3 memory adapter in `SchemaDumperTest` when switching connections (via `ActiveRecord::Base.establish_connection(:arunit2)` in this case), it just removed everything was in `arunit`.