@matthewd 2023-02-27
I haven't looked through this yet, but I'll just note that some of the changes in the previous PR already made me slightly nervous, despite the (clearly valuable) gains... in a quick glance, it seemed like there were at least a few `assert_no_match` anti-assertions that seemed much less protective given a deliberately-constrained input than they'd had before. 🤔