@casperisfine 2023-02-28
> where we can work with a much smaller `:arunit2` connections' tables.

Yeah, it's indeed faster, but do we cover as many combinaisons of types, defaults etc?