@fatkodima 2023-02-28
> Yeah, it's indeed faster, but do we cover as many combinaisons of types, defaults etc?

The only place where it probably became worse, imo, is where we testing something on the whole schema cache dump - https://github.com/rails/rails/pull/47510/files#diff-b525a320ece1314a3a82c4ce58edb382084dc486a12c945e8226f7786c087853R31-R35
Previously, it dumped a lot of tables with lots of different "interesting cases", now only a few. So this specific test case changes can be reverted.

In all other changed places it just tests specific cases (like dumping indexes, or ignoring tables via regexp etc), which we can test on a smaller set of tables, and which was done.

Schema dumper already does dumping of a lot of tables (https://github.com/rails/rails/blob/31ddd41e5876246162a978b7a68bfeacae399d7d/activerecord/test/cases/schema_dumper_test.rb#L48), so we are good there, I think.