@casperisfine 2023-02-28
> So this specific test case changes can be reverted.

I think it would make sense. Other than that, LGTM.