@yahonda 2023-02-28
IMO, this pull request should not be merged until https://github.com/rails/rails/pull/47499#issuecomment-1446691989 is resolved.