@fatkodima 2023-02-28
That should be fixed by https://github.com/rails/rails/pull/47537, I hope 🤞. I am looking at the CI for the merged commits into `main` since then to see if it has not appeared again.