@yahonda 2023-03-01
Thanks for the investigation. Although I also do not know why the logger is related to sqlite3_mem failures. I am happy to see that sqlite3_mem tests are running fine.